# frontend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Frametric API",
    "description": "Cinematic Analytics Platform — REST API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5168"
    }
  ],
  "paths": {
...
    "/api/v1/custom-lists/{id}/movies/{movieId}": {
      "delete": {
        "tags": [
          "CustomLists"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/Movies/{id}": {
      "get": {
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MovieDetailsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieDetailsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Movies/{id}/collection": {
      "get": {
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TmdbCollectionResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TmdbCollectionResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TmdbCollectionResultDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Movies/enrich-from-tmdb": {
      "post": {
        "tags": [
          "Movies"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrichMovieRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrichMovieRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EnrichMovieRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MovieSimpleDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieSimpleDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieSimpleDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Movies/{id}/log": {
      "post": {
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogMovieWatchRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LogMovieWatchRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LogMovieWatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Movies/{id}/log/{entryId}": {
      "delete": {
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Movies/{id}/watchlist": {
      "post": {
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActorSimpleDto": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "DirectorSimpleDto": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "EnrichMovieRequest": {
        "type": "object",
        "properties": {
          "tmdbId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "GenreSimpleDto": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "LogMovieWatchRequest": {
        "type": "object",
        "properties": {
          "dateWatched": {
            "type": "string",
            "format": "date"
          },
          "rating": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "isRewatch": {
            "type": "boolean"
          }
        }
      },
...
      "MovieDetailsDto": {
        "required": [
          "id",
          "title",
          "releaseYear",
          "runtimeMinutes",
          "posterUrl",
          "overview",
          "tmdbRating",
          "userAverageScore",
          "genres",
          "directors",
          "actors",
          "diaryEntries"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "releaseYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "runtimeMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "posterUrl": {
            "type": "string",
            "nullable": true
          },
          "overview": {
            "type": "string",
            "nullable": true
          },
          "tmdbRating": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "userAverageScore": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenreSimpleDto"
            }
          },
          "directors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DirectorSimpleDto"
            }
          },
          "actors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActorSimpleDto"
            }
          },
          "diaryEntries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovieDiaryEntryDto"
            }
          },
          "isWatched": {
            "type": "boolean",
            "default": false
          },
          "isInWatchlist": {
            "type": "boolean",
            "default": false
          },
          "tmdbCollectionId": {
            "type": "integer",
            "format": "int32",
            "default": null,
            "nullable": true
          },
          "tmdbCollectionName": {
            "type": "string",
            "default": null,
            "nullable": true
          }
        }
      },
      "MovieDiaryEntryDto": {
        "required": [
          "id",
          "dateWatched",
          "isRewatch",
          "rating"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "dateWatched": {
            "type": "string"
          },
          "isRewatch": {
            "type": "boolean"
          },
          "rating": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        }
      },
      "MovieSimpleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "releaseYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "posterPath": {
            "type": "string",
            "nullable": true
          },
          "isWatched": {
            "type": "boolean"
          },
          "nickname": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "TmdbCollectionPartDto": {
        "required": [
          "id",
          "title",
          "releaseDate",
          "posterUrl",
          "isInDatabase"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "releaseDate": {
            "type": "string",
            "nullable": true
          },
          "posterUrl": {
            "type": "string",
            "nullable": true
          },
          "isInDatabase": {
            "type": "boolean"
          }
        }
      },
      "TmdbCollectionResultDto": {
        "required": [
          "id",
          "name",
          "overview",
          "posterUrl",
          "backdropUrl",
          "parts"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "overview": {
            "type": "string",
            "nullable": true
          },
          "posterUrl": {
            "type": "string",
            "nullable": true
          },
          "backdropUrl": {
            "type": "string",
            "nullable": true
          },
          "parts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TmdbCollectionPartDto"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Paste your JWT access token here (without the 'Bearer ' prefix).",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}